Simplified derivation of Fastcgipp::Sql::Data::Set
The derivation was simplified by reducing the virtual indexing functions
from three to one. This involved the creation of the class
Fastcgipp::Sql::Data::Index which holds all the data that the three
previous functions returned. Since the new class is constructed by a
large set of constructors that accept references to various types within
the derived data set the issue of mismatching types with void pointers
and creating debugging nightmares has been eliminated.