Fixed template code bloat issue in Manager
Because Fastcgipp::Manager was fully templated a large amount of code
was not being compiled into the library and instead compiled into
individual fastcgi++ applications. This was solved by deriving Manager
from a newly created non-template Fastcgipp::ManagerPar class with all
code that does not depend on the template parameter.