Improved bootstrap script
Modified the boostrap script to better automate the date and version
variables used by autotools, the NEWS file and the doxygen doc
generation. This involved using bash commands to automatically detect
and set the date and version. The NEWS and configure.ac files are set
with appropriate information by using an .in system with them and
sedding them in the boostrap file.