Moved SQL facilities out of Fastcgipp namespace
In order to facilitate the use of the SQL facilities outside of
fastcgi++, it has been moved out of the Fastcgipp namespace into it's
own ASql namespace. In addition the Makefile.am for the includes
directory was modified to not install the sql.hpp file in the event that
no SQL engines are discovered in the configure process. The MySQL error
exception was changed to not rely on the fastcgi++ parent exception.