Removed ASql dependence on message.hpp
The callback mechanism of the ASql part now takes an argument for it's own
error exception as opposed to a Fastcgipp::Message object. As a result a
new system of intermediate binding was coded into the asqlbind files.
This adds a small layer to execution of asynchronous queries but allows
the ASql part to be easily broken away from fastcgi++ and used on it's
own.