Reimplemented ASql::Data::SetContainer again
This time I made SetContainer take an actual container type as a
template parameter. This way the user can use whatever container type
they want so long as has a few needed member functions. The class
derives from the type passed to it so it should behave almost exactly
like the container form the users point of view.