Fixed REQUEST_METHOD bug that broke echo.cpp
The initialization of Http::Environment::requestMethod was done using
non-templated string types which broke the echo example as you cannot
output std::string into a wide character stream. This was corrected be
removing dependencies on a separate object to hold the string data all
together and creating a stream inserter for the Http::RequestMethod
enumeration.