Changed response error handling
Request now outputs more details error messages to the error log upon
catching an exception and then outputs an appropriate HTTP error header
to the client. This is kind of broken because it looks like the web
server doesn't actually pass along it's canned response associated with
the error header, just the header itself.

Also implemented an exception to be thrown inside the response for
FastCGI records received out of order.