Completely change ASql queue functionality
As opposed to passing a series of boost::shared_ptrs to a queue function
for asynchronous queries, the data has been assembled into a structure
that transparently manages data lifespans. It also provides a means of
canceling a query should the user be tired of waiting. A cancel can also
happen automatically should the query object go out of scope in the
calling thread.