Finished (more or less) this patch.
Unfortunately, there is no unique character to identify the end of
application/x-www-form-urlencoded data in a POST request. As a result,
if the last record in a POST request is not as expected (i.e.
key=value), there is no way to detect if it is incomplete or malformed.

Aside from missing that particular case, it works.