Add support for querying and fetching of post variables.
Added postVariableExists to test for the existence of a post variable
and postVariableRetrieve to retrieve the value associated with variable
key and cast it to the type desired by the user. More details included
in the doxygen documentation.

One caveat: I could not specialize postVariableRetrieve char and wchar_t
as C++ does not allow partial specialization of function templates. I
had originally overloaded it for each type, but saw little advantage to
doing that instead of just leaving it without specialization. That
option exists though if it's decided later that it's a better approach.

Also, feel free to rename both functions.