Fixed padding bug in Http::base64Decode()
Fastcgipp::Http::base64Decode() was not properly handling the padding at
the end of base64 encoding strings. Extra null characters were being
added when they shouldn't have been.

Along with this fix comes a change to the function. It now returns an
iterator to the last characters written +1. The way it should have
worked from the beginning.

Thanks to Henrique Aparicio for finding this little nugget.