Improved concurrency system in ASql
The ASql system is now fully concurrent. When building a connection
object the amount of threads passed to the constructor allows that
amount of queries to be executing simultaneously. The effectively
creates separate queues for each thread. Queueing up a query always
places it in the smaller queue. Obviously transactions are guaranteed to
always be placed in the same queues.