Eliminated ineffecient POST handling
The Http::Environment class was filling its POST data in a very stupid
way. It was constantly re-copying data around and re-parsing it. The
system now simple waits until all POST data has been recieved into a
single buffer before parsing it. If there is any content beyond the
content length passed from the client it is truncated. Special thanks to
Smeesters Olivier for catching this.