Accommodating Chrome/Safari file multipart
Both Chrome and Safari have taken to adding an additional "\r\n" to the
end of multipart encoded POST files. The system now checks for said
character sequence at the end of the file and truncates if it exists.

This of course breaks POSTing of files with Mozilla/IE that legitimately
end with a "\r\n" but I'm not sure how else to proceed.